Move PresentationalSearch back to the container
